% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_api.R
\name{query_api}
\alias{query_api}
\title{Query a Paginated App Store Connect API Endpoint}
\usage{
query_api(verb, api_url, jwt, ...)
}
\arguments{
\item{verb}{A verb indicating the HTTP method to use (must be from \code{{httr}}).}

\item{api_url}{A string representing the initial API endpoint URL to query.}

\item{jwt}{A string containing the JSON Web Token (JWT) used for authentication.}

\item{...}{arguments passed to \code{verb}}
}
\value{
A list containing all results retrieved from the API across all pages.
If an error occurs during any request, \code{NULL} is returned, and the error message is printed.
}
\description{
This function queries a specified App Store Connect API endpoint, handling authentication
with a JSON Web Token (JWT) and iterating through paginated results until all data is retrieved.
}
\details{
\itemize{
\item The function sends a \code{verb} request to the specified \code{api_url}.
\item Authentication is handled by attaching the JWT in the \code{Authorization} header.
\item If the response includes a \code{meta$paging$next} field, the function retrieves additional pages.
\item Results from all pages are combined into a single list and returned.
}
}
\examples{
\dontrun{
jwt <- generate_jwt(
  issuer_id = '1234',
  key_id = '1234',
  private_key_path = "1234"
)
query_api(
  httr::GET,
  "https://api.appstoreconnect.apple.com/v1/apps",
   jwt
)
}
}
